Cleanup @vanilla/utils imports and separate tests into different files
