+++ packages/vanilla-utils/package.json
@@ -1,6 +1,9 @@
 {
     "name": "@vanilla/utils",
-    "main": "./src/index.ts",
-    "module": "./src/index.ts",
-    "version": "3.0.0"
+    "main": "./index.ts",
+    "module": "./index.ts",
+    "version": "3.0.0",
+    "devDependencies": {
+        "@vanilla/tsconfig": "*"
+    }
 }
