--- packages/vanilla-utils/src/index.ts
@@ -1,15 +1,0 @@
-/**
- * General utility functions.
- * This file should have NO external dependencies other than javascript.
- *
- * @copyright 2009-2019 Vanilla Forums Inc.
- * @license GPL-2.0-only
- */
-
-export * from "./debugUtils";
-export * from "./fileUtils";
-export * from "./mathUtils";
-export * from "./promiseUtils";
-export * from "./stringUtils";
-export * from "./systemUtils";
-export * from "./typeUtils";
