+++ packages/vanilla-utils/tsconfig.json
@@ -1,0 +1,4 @@
+{
+    "extends": "@vanilla/tsconfig/base",
+    "compilerOptions": {}
+}
